script/e2eTest/pyproject.toml
[project]
name = "e2etest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "boto3>=1.37.38",
    "click>=8.1.8",
    "pyyaml>=6.0.2",
]
